ORC-1221: Move NullKeyProvider to org.apache.orc.impl
